+++ src/rai_bench/rai_bench/manipulation_o3de/interfaces.py
@@ -17,11 +17,11 @@ from abc import ABC, abstractmethod
 from collections import defaultdict
 from typing import Dict, List, Set, Tuple, TypeVar, Union
 
+from rai.types import Pose
 from rclpy.impl.rcutils_logger import RcutilsLogger
 
 from rai_sim.simulation_bridge import (
     Entity,
-    Pose,
     SimulationBridge,
     SimulationConfig,
     SimulationConfigT,
