--- src/rai_core/rai/types/gazebo.py
@@ -1,26 +1,0 @@
-# Copyright (C) 2025 Robotec.AI
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#         http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-from pydantic import Field
-
-from .base import Ros2BaseModel
-from .geometry import Pose
-
-
-class SpawnEntityService(Ros2BaseModel):
-    _prefix: str = "gazebo_msgs/srv"
-    name: str
-    robot_namespace: str
-    reference_frame: str
-    initial_pose: Pose
-    xml: str = Field(default="")
